feat(safety): move code
